Conflicts:
	code/modules/mob/living/simple_animal/bot/ed209bot.dm

+++ code/game/objects/structures/morgue.dm
@@ -16,7 +16,7 @@
 	icon = 'icons/obj/stationobjs.dmi'
 	icon_state = "morgue1"
 	density = 1
-	anchored = 1.0
+	anchored = 1
 
 	var/obj/structure/tray/connected = null
 	var/locked = 0
@@ -30,7 +30,7 @@
 	if(connected)
 		qdel(connected)
 		connected = null
-	..()
+	return ..()
 
 /obj/structure/bodycontainer/on_log()
 	update_icon()
@@ -41,10 +41,16 @@
 /obj/structure/bodycontainer/alter_health()
 	return src.loc
 
-/obj/structure/bodycontainer/attack_paw(mob/user as mob)
+/obj/structure/bodycontainer/relaymove(mob/user)
+	if(user.stat || !isturf(loc))
+		return
+	if(!open())
+		open()
+
+/obj/structure/bodycontainer/attack_paw(mob/user)
 	return src.attack_hand(user)
 
-/obj/structure/bodycontainer/attack_hand(mob/user as mob)
+/obj/structure/bodycontainer/attack_hand(mob/user)
 	if(locked)
 		user << "<span class='danger'>It's locked.</span>"
 		return
@@ -57,7 +63,7 @@
 		close()
 	add_fingerprint(user)
 
-/obj/structure/bodycontainer/attackby(P as obj, mob/user as mob, params)
+/obj/structure/bodycontainer/attackby(obj/P, mob/user, params)
 	if (istype(P, /obj/item/weapon/pen))
 		var/t = stripped_input(user, "What would you like the label to be?", text("[]", name), null)
 		if (user.get_active_hand() != P)
@@ -135,7 +141,7 @@ var/global/list/crematoriums = new/list()
 
 /obj/structure/bodycontainer/crematorium/Destroy()
 	crematoriums.Remove(src)
-	..()
+	return ..()
 
 /obj/structure/bodycontainer/crematorium/New()
 	connected = new/obj/structure/tray/c_tray(src)
@@ -159,7 +165,7 @@ var/global/list/crematoriums = new/list()
 
 	return
 
-/obj/structure/bodycontainer/crematorium/proc/cremate(mob/user as mob)
+/obj/structure/bodycontainer/crematorium/proc/cremate(mob/user)
 	if(locked)
 		return //don't let you cremate something twice or w/e
 
@@ -174,15 +180,17 @@ var/global/list/crematoriums = new/list()
 		update_icon()
 
 		for(var/mob/living/M in contents)
-			if (M.stat!=2)
+			if (M.stat != DEAD)
 				M.emote("scream")
-			//Logging for this causes runtimes resulting in the cremator locking up. Commenting it out until that's figured out.
-			//M.attack_log += "\[[time_stamp()]\] Has been cremated by <b>[user]/[user.ckey]</b>" //No point in this when the mob's about to be deleted
-			user.attack_log +="\[[time_stamp()]\] Cremated <b>[M]/[M.ckey]</b>"
-			log_attack("\[[time_stamp()]\] <b>[user]/[user.ckey]</b> cremated <b>[M]/[M.ckey]</b>")
+			if(user)
+				user.attack_log +="\[[time_stamp()]\] Cremated <b>[M]/[M.ckey]</b>"
+				log_attack("\[[time_stamp()]\] <b>[user]/[user.ckey]</b> cremated <b>[M]/[M.ckey]</b>")
+			else
+				log_attack("\[[time_stamp()]\] <b>UNKNOWN</b> cremated <b>[M]/[M.ckey]</b>")
 			M.death(1)
-			M.ghostize()
-			qdel(M)
+			if(M) //some animals get automatically deleted on death.
+				M.ghostize()
+				qdel(M)
 
 		for(var/obj/O in contents) //obj instead of obj/item so that bodybags and ashes get destroyed. We dont want tons and tons of ash piling up
 			if(O != connected) //Creamtorium does not burn hot enough to destroy the tray
@@ -193,26 +201,6 @@ var/global/list/crematoriums = new/list()
 		locked = 0
 		update_icon()
 		playsound(src.loc, 'sound/machines/ding.ogg', 50, 1) //you horrible people
-
-/*
-Crematorium Switch
-*/
-/obj/machinery/crema_switch/attack_hand(mob/user as mob)
-	if(src.allowed(usr))
-		for (var/obj/structure/bodycontainer/crematorium/C in crematoriums)
-			if (C.id != id)
-				continue
-
-			C.cremate(user)
-	else
-		usr << "<span class='danger'>Access denied.</span>"
-	return
-
-/obj/machinery/crema_switch/attackby(obj/item/W as obj, mob/user as mob, params)
-	if(W.GetID())
-		attack_hand(user)
-	else
-		return ..()
 
 
 /*
@@ -225,27 +213,27 @@ Crematorium Switch
 	density = 1
 	layer = 2.9
 	var/obj/structure/bodycontainer/connected = null
-	anchored = 1.0
-	throwpass = 1
+	anchored = 1
+	pass_flags = LETPASSTHROW
 
 /obj/structure/tray/Destroy()
 	if(connected)
 		connected.connected = null
 		connected.update_icon()
 		connected = null
-	..()
+	return ..()
 
-/obj/structure/tray/attack_paw(mob/user as mob)
+/obj/structure/tray/attack_paw(mob/user)
 	return src.attack_hand(user)
 
-/obj/structure/tray/attack_hand(mob/user as mob)
+/obj/structure/tray/attack_hand(mob/user)
 	if (src.connected)
 		connected.close()
 		add_fingerprint(user)
 	else
 		user << "<span class='warning'>That's not connected to anything!</span>"
 
-/obj/structure/tray/MouseDrop_T(atom/movable/O as mob|obj, mob/user as mob)
+/obj/structure/tray/MouseDrop_T(atom/movable/O as mob|obj, mob/user)
 	if(!istype(O, /atom/movable) || O.anchored || !Adjacent(user) || !user.Adjacent(O) || O.loc == user)
 		return
 	if(!ismob(O))
@@ -278,3 +266,12 @@ Crematorium Switch
 	desc = "Apply corpse before closing."
 	icon_state = "morguet"
 
+/obj/structure/tray/m_tray/CanPass(atom/movable/mover, turf/target, height=0)
+	if(height==0) return 1
+
+	if(istype(mover) && mover.checkpass(PASSTABLE))
+		return 1
+	if(locate(/obj/structure/table) in get_turf(mover))
+		return 1
+	else
+		return 0
