+++ code/modules/food&drinks/kitchen machinery/monkeyrecycler.dm
@@ -1,6 +1,6 @@
 /obj/machinery/monkey_recycler
 	name = "monkey recycler"
-	desc = "A machine used for recycling dead monkeys into monkey cubes. It requires 5 monkeys per cube."
+	desc = "A machine used for recycling dead monkeys into monkey cubes. It currently produces 1 cube for every 5 monkeys inserted." // except it literally never does
 	icon = 'icons/obj/kitchen.dmi'
 	icon_state = "grinder"
 	layer = 2.9
@@ -31,8 +31,9 @@
 		cubes_made = M.rating
 	cube_production = cubes_made
 	required_grind = req_grind
+	src.desc = "A machine used for recycling dead monkeys into monkey cubes. It currently produces [cubes_made] cube(s) for every [required_grind] monkey(s) inserted."
 
-/obj/machinery/monkey_recycler/attackby(var/obj/item/O as obj, var/mob/user as mob, params)
+/obj/machinery/monkey_recycler/attackby(obj/item/O, mob/user, params)
 	if(default_deconstruction_screwdriver(user, "grinder_open", "grinder", O))
 		return
 
@@ -55,29 +56,32 @@
 		if(!user.Adjacent(G.affecting))
 			return
 		var/grabbed = G.affecting
-		if(istype(grabbed, /mob/living/carbon/monkey))
+		if(ismonkey(grabbed))
 			var/mob/living/carbon/monkey/target = grabbed
 			if(target.stat == 0)
-				user << "<span class='danger'>The monkey is struggling far too much to put it in the recycler.</span>"
-			else
-				if(!user.drop_item())
-					return
-				qdel(target)
-				user << "<span class='notice'>You stuff the monkey in the machine.</span>"
-				playsound(src.loc, 'sound/machines/juicer.ogg', 50, 1)
-				var/offset = prob(50) ? -2 : 2
-				animate(src, pixel_x = pixel_x + offset, time = 0.2, loop = 200) //start shaking
-				use_power(500)
-				src.grinded++
-				sleep(50)
-				pixel_x = initial(pixel_x) //return to its spot after shaking
-				user << "<span class='notice'>The machine now has [grinded] monkey\s worth of material stored.</span>"
+				user << "<span class='warning'>The monkey is struggling far too much to put it in the recycler.</span>"
+				return
+			if(target.buckled || target.buckled_mob)
+				user << "<span class='warning'>The monkey is attached to something.</span>"
+				return
+			if(!user.drop_item())
+				return
+			qdel(target)
+			user << "<span class='notice'>You stuff the monkey into the machine.</span>"
+			playsound(src.loc, 'sound/machines/juicer.ogg', 50, 1)
+			var/offset = prob(50) ? -2 : 2
+			animate(src, pixel_x = pixel_x + offset, time = 0.2, loop = 200) //start shaking
+			use_power(500)
+			grinded++
+			sleep(50)
+			pixel_x = initial(pixel_x) //return to its spot after shaking
+			user << "<span class='notice'>The machine now has [grinded] monkey\s worth of material stored.</span>"
 
 		else
 			user << "<span class='danger'>The machine only accepts monkeys!</span>"
 	return
 
-/obj/machinery/monkey_recycler/attack_hand(var/mob/user as mob)
+/obj/machinery/monkey_recycler/attack_hand(mob/user)
 	if (src.stat != 0) //NOPOWER etc
 		return
 	if(grinded >= required_grind)
